Download games and workshop items to a separate dir
It was previously mistakenly thought that steamcmd would download
"app_update" files to a subdirectory of the force_install_dir.
This is not the case. And because force_install_dir is discouraged to
be ran after login, steamcmd needs to run multiple times and a separate
workshop items install dir is needed.